requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.6.0
PyMuPDF>=1.24.0
langgraph>=0.0.30
langchain>=0.1.12
langchain-core>=0.1.33
python-multipart>=0.0.9
requests>=2.31.0
python-dotenv>=1.0.1
google-generativeai>=0.4.0
Pillow>=10.0.0
opencv-python-headless>=4.9.0
